Merge in updates from VectorCAST jenkins execution plugin master

+++ src/main/resources/scripts/generate-results.py
@@ -21,6 +21,10 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #
+
+from __future__ import print_function
+from __future__ import absolute_import
+
 import os
 import sys
 import argparse
@@ -30,13 +34,17 @@ import glob
 import subprocess
 import time
 import traceback
+import parse_traceback
+import tee_print
 
 # adding path
 jenkinsScriptHome = os.getenv("WORKSPACE") + os.sep + "vc_scripts"
 python_path_updates = jenkinsScriptHome
 sys.path.append(python_path_updates)
-python_path_updates += os.sep + "vpython-addons"
-sys.path.append(python_path_updates)
+# needed because vc18 vpython does not have bs4 package
+if sys.version_info[0] < 3:
+    python_path_updates += os.sep + 'vpython-addons'
+    sys.path.append(python_path_updates)
 
 import tcmr2csv
 import vcastcsv2jenkins
@@ -55,71 +63,31 @@ except:
     
 #global variables
 global verbose
+global print_exc
 global wait_time
 global wait_loops
 
 verbose = False
+print_exc = False
 
-#
-# Internal - jUnit works best with one overall file for all test results
-#
-def writeJunitFinalCombinedTestResults(manageProjectName):
-    fname = manageProjectName
-    totalFailed = 0
-    
-    f=open("xml_data/test_results_"+fname+"_combined_final.xml","w")
-    f.write("<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n")
-    f.write("<testsuites>\n")
+import getjobs
 
-    for testResults in glob.glob('xml_data/test_results*.xml'):
-        if testResults.endswith("_combined_final.xml"):
-            continue
-        if not testResults.endswith("_combined.xml"):
-            os.remove(testResults)
-            continue
-            
-        data = open(testResults,"r").readlines()[2:-1]
-        for line in data:
-            if "<testsuite errors=" in line:
-                totalFailed += int(line.split("\"")[5])
-        os.remove(testResults)
-        wrData = "".join(data)
-        f.write(wrData)
-     
-    f.write("\n</testsuites>\n")
-    f.close()
-    
-    open("unit_test_fail_count.txt","w").write(str(totalFailed))
+enabledEnvironmentArray = []
 
-#
-# Internal - jUnit works best with one overall file for all test results
-#
-def writeJunitCombinedTestResults(manageProjectName,orig_level,envName):
-    fname = manageProjectName
-    if orig_level:
-        level = orig_level.split("/")
-        if len(level) == 2:
-            # Level does not include source and platform
-            jobName = level[0] + "_" + level[1].rstrip()
-        else:
-            # Level includes source and platform
-            jobName = level[2] + "_" + level[3].rstrip()
-        fname = manageProjectName + "_" + jobName + "_" + envName
-        
-    f=open("xml_data/test_results_"+fname+"_combined.xml","w")
-    f.write("<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n")
-    f.write("<testsuites>\n")
+def getEnabledEnvironments(MPname):
+    output = getjobs.printEnvironmentInfo(MPname, False)
 
-    for testResults in glob.glob('xml_data/test_results*.xml'):
-        if  "combined" in testResults:
-            continue
-        data = open(testResults,"r").readlines()[2:-1]
-        os.remove(testResults)
-        wrData =  "".join(data)
-        f.write(wrData)
-     
-    f.write("\n</testsuites>\n")
-    f.close()
+    for line in output.split("\n"):
+        if line.strip():
+            compiler, testsuite, environment = line.split()
+            enabledEnvironmentArray.append([compiler, testsuite, environment])
+                       
+def environmentEnabled(comp,ts,env):
+    for c,t,e in enabledEnvironmentArray:
+        if comp == c and ts == t and env == e:
+            return True
+    print(comp + "/" + ts + "/" + env + ": Disabled")
+    return False 
 
 def runManageWithWait(command_line, silent=False):
     global verbose
@@ -134,8 +102,8 @@ def checkUseNewReportsAndAPI():
     if os.environ.get("VCAST_REPORT_ENGINE", "") == "LEGACY":
         # Using legacy reporting with new reports - fall back to parsing html report
         if verbose:
-            print("Reports forced to be old/legacy, so use them")
-        return False
+            print("VectorCAST/Execution ignoring LEGACY VCAST_REPORT_ENGINE.")
+
     # Look for existence of file that only exists in distribution with the new reports
     check_file = os.path.join(os.environ.get('VECTORCAST_DIR'),
                              "python",
@@ -193,6 +161,8 @@ def getManageEnvs(FullManageProjectName):
             #rare case where there's a problem with the environment
             if build_dir == "":
                 continue
+            if not environmentEnabled(compiler,testsuite,env_name):
+                continue
             build_dir_number = build_dir.split("/")[-1]
             level = compiler + "/" + testsuite + "/" + env_name # env_name.upper()
             entry = {}
@@ -217,7 +187,7 @@ def delete_file(filename):
     if os.path.exists(filename):
         os.remove(filename)
         
-def genDataApiReports(entry, jUnit, cbtDict):
+def genDataApiReports(FullManageProjectName, entry, cbtDict):
     xml_file = ""
     
     try:
@@ -233,7 +203,8 @@ def genDataApiReports(entry, jUnit, cbtDict):
         xmlUnitReportName = os.getcwd() + os.sep + "xml_data" + os.sep + "test_results_" + level + "_" + env + ".xml"
         xmlCoverReportName = os.getcwd() + os.sep + "xml_data" + os.sep + "coverage_results_" + level + "_" + env + ".xml"
 
-        xml_file = GenerateXml(entry["build_dir"],
+        xml_file = GenerateXml(FullManageProjectName,
+                               entry["build_dir"],
                                entry["env"],entry["compiler"],entry["testsuite"],
                                xmlCoverReportName,
                                jenkins_name,
@@ -241,7 +212,6 @@ def genDataApiReports(entry, jUnit, cbtDict):
                                jenkins_link,
                                jobNameDotted, 
                                verbose, 
-                               jUnit,
                                cbtDict)
                                
         if xml_file.api != None:
@@ -255,12 +225,16 @@ def genDataApiReports(entry, jUnit, cbtDict):
         else:
             print("   Skipping environment: " + jobNameDotted)
             
-    except Exception as e:
-        print("ERROR: failed to generate XML reports using vpython and the Data API for ", entry["compiler"] + "_" + entry["testsuite"] + "_" + entry["env"], "in directory", entry["build_dir"])
-        traceback.print_exc()
-        
-    return xml_file
     
+    except Exception as e:
+        parse_traceback.parse(traceback.format_exc(), print_exc, entry["compiler"] , entry["testsuite"],  entry["env"], entry["build_dir"])
+        
+    try:       
+        return xml_file.failed_count
+    except:
+        return 0
+        
+        
 def fixup_css(report_name):
     # Needed for VC19 and VC19 SP1.
     # From VC19 SP2 onwards a new option VCAST_RPTS_SELF_CONTAINED is used instead
@@ -306,8 +280,10 @@ def generateCoverReport(path, env, level ):
 
         fixup_css(report_name)
     except Exception as e:
-        print("   *Problem generating custom report for " + env + ": ")
-        traceback.print_exc()
+        build_dir = path.replace("\\","/")
+        build_dir = build_dir.rsplit("/",1)[0]
+
+        parse_traceback.parse(traceback.format_exc(), print_exc, level.split("_")[0] , level.split("_")[2], env, build_dir)
 
 def generateUTReport(path, env, level): 
     global verbose
@@ -322,9 +298,11 @@ def generateUTReport(path, env, level):
         api.report(report_type="FULL_REPORT", formats=["HTML"], output_file=report_name)
         fixup_css(report_name)
     except Exception as e:
-        print("   *Problem generating custom report for " + env + ".")
-        traceback.print_exc()
+        build_dir = path.replace("\\","/")
+        build_dir = build_dir.rsplit("/",1)[0]
 
+        parse_traceback.parse(traceback.format_exc(), print_exc, level.split("_")[0] , level.split("_")[2], env, build_dir)
+        
 def generateIndividualReports(entry, envName):
     global verbose
 
@@ -342,25 +320,58 @@ def generateIndividualReports(entry, envName):
         elif os.path.exists(unit_path):
             generateUTReport(unit_path , env, level)                
 
-def useNewAPI(manageEnvs, level, envName, jUnit, cbtDict):
+def useNewAPI(FullManageProjectName, manageEnvs, level, envName, cbtDict):
+    failed_count = 0 
         
     for currentEnv in manageEnvs:
         if envName == None:
-            genDataApiReports(manageEnvs[currentEnv], jUnit, cbtDict)
+            failed_count += genDataApiReports(FullManageProjectName, manageEnvs[currentEnv],  cbtDict)
             generateIndividualReports(manageEnvs[currentEnv], envName)
             
         elif manageEnvs[currentEnv]["env"].upper() == envName.upper(): 
             env_level = manageEnvs[currentEnv]["compiler"] + "/" + manageEnvs[currentEnv]["testsuite"]
             
             if env_level.upper() == level.upper():
-                genDataApiReports(manageEnvs[currentEnv], jUnit, cbtDict)
+                failed_count += genDataApiReports(FullManageProjectName, manageEnvs[currentEnv], cbtDict)
                 generateIndividualReports(manageEnvs[currentEnv], envName)
+    f = open("unit_test_fail_count.txt","w")
+    f.write(str(failed_count))
+    f.close()
 
+
+def cleanupDirectory(path, teePrint):
+    # if the path exists, try to delete it
+    if os.path.isdir(path):
+        try:
+            shutil.rmtree(path)
+        except:
+            # if there was an error removing the directory...delete all the files
+            print("Error removing directory: " + path)
+            for file in glob.glob(path + "/*.*"):
+                try:
+                    os.remove(file);
+                except:
+                    teePrint.teePrint("   *INFO: File System Error removing file after failed to remove directory: " + path + "/" + file + ".  Check console for environment build/execution errors")
+                    if print_exc:  traceback.print_exc()
+            pass
+
+    # we should either have an empty directory or no directory
+    if not os.path.isdir(path):
+        try:
+            os.mkdir(path)
+        except:
+            teePrint.teePrint("   *INFO: File System Error creating directory: " + path + ".  Check console for environment build/execution errors")
+            if print_exc:  traceback.print_exc()
+
+
+def cleanupOldBuilds(teePrint):
+    for path in ["xml_data","management","execution"]:
+        cleanupDirectory(path, teePrint)
 
 # build the Test Case Management Report for Manage Project
 # envName and level only supplied when doing reports for a sub-project
 # of a multi-job
-def buildReports(FullManageProjectName = None, level = None, envName = None, generate_individual_reports = True, timing = False, jUnit = True, cbtDict = None):
+def buildReports(FullManageProjectName = None, level = None, envName = None, generate_individual_reports = True, timing = False, cbtDict = None):
 
     if timing:
         print("Start: " + str(time.time()))
@@ -382,43 +393,24 @@ def buildReports(FullManageProjectName = None, level = None, envName = None, gen
     if timing:
         print("Version Check: " + str(time.time()))
 
-    # cleaning up old builds
-    for path in ["xml_data","management","execution"]:
-        # if the path exists, try to delete it
-        if os.path.isdir(path):
-            try:
-                shutil.rmtree(path)
-            except:
-                # if there was an error removing the directory...delete all the files
-                print("Error removing directory: " + path)
-                for file in glob.glob(path + "/*.*"):
-                    try:
-                        os.remove(file);
-                    except:
-                        print("Error removing file after failed to remove directory: " + path + "/" + file)
-                pass
-                
-        # we should either have an empty directory or no directory
-        if not os.path.isdir(path):
-            try:
-                os.mkdir(path)
-            except:
-                print("Error creating directory: " + path)
-            
+    with tee_print.TeePrint() as teePrint:
+        cleanupOldBuilds(teePrint)
+
+
     for file in glob.glob("*.csv"):
         try:
             os.remove(file);
             if verbose:
                 print("Removing file: " + file)
         except Exception as e:
-            print("Error removing " + file)
-            print(e)
-    
+            teePrint.teePrint("   *INFO: File System Error removing " + file + ".  Check console for environment build/execution errors")
+            if print_exc:  traceback.print_exc()
+   
     
     ### Using new data API - 2019 and beyond
     if timing:
         print("Cleanup: " + str(time.time()))
-    if useNewReport:
+    if useNewReport and not legacy:
 
         try:
             shutil.rmtree("execution") 
@@ -428,7 +420,7 @@ def buildReports(FullManageProjectName = None, level = None, envName = None, gen
         if timing:
             print("Using DataAPI for reporting")
             print("Get Info: " + str(time.time()))
-        useNewAPI(manageEnvs, level, envName, jUnit, cbtDict)
+        useNewAPI(FullManageProjectName, manageEnvs, level, envName, cbtDict)
         if timing:
             print("XML and Individual reports: " + str(time.time()))
 
@@ -455,12 +447,12 @@ def buildReports(FullManageProjectName = None, level = None, envName = None, gen
         # capture the output of the manage call
         out_mgt = runManageWithWait(callStr)
 
-        missing = False
+        coverProjectInManageProject = False
         if "database missing or inaccessible" in out_mgt:
-            missing = True
+            coverProjectInManageProject = True
         elif re.search('Environment directory.*is missing', out_mgt):
-            missing = True
-        if missing:
+            coverProjectInManageProject = True
+        if coverProjectInManageProject:
             callStr = callStr.replace("report custom","cover report")
             print(callStr)
             out_mgt2 = runManageWithWait(callStr)
@@ -548,7 +540,7 @@ def buildReports(FullManageProjectName = None, level = None, envName = None, gen
                     # Create the test_results_ and coverage_results_ csv files
                     testResultName, coverageResultsName = tcmr2csv.run(reportName, level, version)
 
-                    vcastcsv2jenkins.run  (test = testResultName, coverage = coverageResultsName, useExecRpt=generate_individual_reports, version=version, junit=junit)
+                    vcastcsv2jenkins.run  (test = testResultName, coverage = coverageResultsName, useExecRpt=generate_individual_reports, version=version)
 
                     if envName:
                         adjustedReportName = "management" + os.sep + jobName + "_" + envName + ".html"
@@ -559,6 +551,24 @@ def buildReports(FullManageProjectName = None, level = None, envName = None, gen
                 copyList.append([reportName,adjustedReportName])
                 # Reset env
                 env = None
+        
+        if coverProjectInManageProject:
+            generate_qa_results_xml.genQATestResults(FullManageProjectName,saved_level,saved_envName)
+            
+        failed_count = 0
+        try:
+            for file in glob.glob("xml_data/test_results_*.xml"):
+                lines = open(file,"r").readlines()
+                for line in lines:
+                    if "failures" in line:
+                        failed_count += int(line.split("\"")[5])
+                        break
+        except:
+            teePrint.teePrint ("   *INFO: Problem parsing test results file for unit testcase failure count: " + file)
+            if print_exc:  traceback.print_exc()
+        f = open("unit_test_fail_count.txt","w")
+        f.write(str(failed_count))
+        f.close()
                 
 
         for file in copyList:
@@ -571,14 +581,11 @@ def buildReports(FullManageProjectName = None, level = None, envName = None, gen
     if timing:
         print("QA Results reports: " + str(time.time()))
             
-    if junit:   
-        if verbose:
-            print("Writing combined test data for JUnit")
-        generate_qa_results_xml.genQATestResults(FullManageProjectName, saved_level, saved_envName)
-        writeJunitCombinedTestResults(manageProjectName,saved_level,saved_envName)
 
     if timing:
         print("Complete: " + str(time.time()))
+        
+        
         
 if __name__ == '__main__':
 
@@ -591,17 +598,15 @@ if __name__ == '__main__':
     parser.add_argument('--wait_time',   help='Time (in seconds) to wait between execution attempts', type=int, default=30)
     parser.add_argument('--wait_loops',   help='Number of times to retry execution', type=int, default=1)
     parser.add_argument('--timing',   help='Display timing information for report generation', action="store_true")
-    parser.add_argument('--junit',   help='Output test resutls in junit format', action="store_true")
+    parser.add_argument('--junit',   help='Output test resutls in JUnit format', action="store_true")
+    parser.add_argument('--print_exc',   help='Output test resutls in JUnit format', action="store_true")
     parser.add_argument('--api',   help='Unused', type=int)
-    parser.add_argument('--final',   help='Write Final JUnit Test Results file',  action="store_true")
+
+    parser.add_argument('--legacy',   help='Force legacy reports for testing only', action="store_true", default = False)
     parser.add_argument('--buildlog',   help='Build Log for CBT Statitics')
 
     args = parser.parse_args()
     
-    if args.final:
-        generate_qa_results_xml.genQATestResults(args.ManageProject)
-        writeJunitFinalCombinedTestResults(os.path.basename(args.ManageProject))
-        sys.exit(0)
     try:
         if "19.sp1" in open(os.path.join(os.environ['VECTORCAST_DIR'],"DATA/tools_version.txt").read):
             # custom report patch for SP1 problem - should be fixed in future release      
@@ -617,6 +622,8 @@ if __name__ == '__main__':
 
     if args.verbose:
         verbose = True
+    if args.print_exc:
+        print_exc = True
     wait_time = args.wait_time
     wait_loops = args.wait_loops
 
@@ -633,21 +640,25 @@ if __name__ == '__main__':
     if args.junit:
         junit = True
     else:
-        junit = False
+        print ("Test results reporting has been migrated to JUnit.  If you are using older xUnit plugin with Single Jobs, please switch to using JUnit.  If you need assistance with that, contact support@us.vector.com")
+        junit = True
         
-    if args.buildlog:
+    if args.buildlog and os.path.exists(args.buildlog):
         buildLogData = open(args.buildlog,"r").readlines()
-        cbt = ParseConsoleForCBT()
+        cbt = ParseConsoleForCBT(verbose)
         cbtDict = cbt.parse(buildLogData)
         
     else:
         cbtDict = None
-
-
+        
+    getEnabledEnvironments(args.ManageProject)
+    
     # Used for pre VC19
     os.environ['VCAST_RPTS_PRETTY_PRINT_HTML'] = 'FALSE'
     # Used for VC19 SP2 onwards
     os.environ['VCAST_RPTS_SELF_CONTAINED'] = 'FALSE'
 
-    buildReports(args.ManageProject,args.level,args.environment,dont_generate_individual_reports, timing, junit, cbtDict)
+    legacy = args.legacy
+        
+    buildReports(args.ManageProject,args.level,args.environment,dont_generate_individual_reports, timing, cbtDict)
